requirements.txt
pygame==2.6.1
PyOpenGL>=3.1.4
PyOpenGL_accelerate>=3.1.4
numpy>=1.21.0
matplotlib>=3.5.0
